refactor mapkit implementation.
