+++ SwiftUI_Templates/MapSearchView/MapViewContainer.swift
@@ -1,0 +1,37 @@
+//
+//  MapViewContainer.swift
+//  SwiftUI_Templates
+//
+//  Created by Hui, Malachi | DCMS on 2020/05/16.
+//  Copyright © 2020 Hui, Malachi | DCMS. All rights reserved.
+//
+
+import SwiftUI
+import MapKit
+
+struct MapViewContainer: UIViewRepresentable {
+    var annotations = [MKPointAnnotation]()
+    let mapView = MKMapView()
+    
+    func makeUIView(context: UIViewRepresentableContext<MapViewContainer>) -> MKMapView {
+        setupRegionForMap()
+        return mapView
+    }
+    
+    fileprivate func setupRegionForMap() {
+        //San Francisco Location
+        let centerCoordinate = CLLocationCoordinate2D(latitude: 37.7666, longitude: -122.427290)
+        let span = MKCoordinateSpan(latitudeDelta: 0.1, longitudeDelta: 0.1)
+        let region = MKCoordinateRegion(center: centerCoordinate, span: span)
+        mapView.setRegion(region, animated: true)
+    }
+    
+    func updateUIView(_ uiView: MKMapView, context: UIViewRepresentableContext<MapViewContainer>) {
+        uiView.removeAnnotations(uiView.annotations)
+        uiView.addAnnotations(annotations)
+        uiView.showAnnotations(uiView.annotations, animated: false)
+    }
+    
+    typealias UIViewType = MKMapView
+}
+
